feat: replace Counter with CountdownTimer component and restructure Manifesto component
