+++ src/components/Manifesto/manifesto.css
