+++ src/components/Manifesto/manifesto.tsx
